pyproject.toml
[tool.poetry]
name = "RAG-chat-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
pinecone-client = "3.0.0.dev8"
langchain-community = "0.0.32"
cohere = "^5.3.3"
openai = "^1.23.4"
python-dotenv = "^1.0.1"
streamlit = "^1.33.0"
fastapi = "^0.110.2"
langchain = "^0.1.16"
langchain-openai = "^0.1.3"
langchainhub = "^0.1.15"
langsmith = "^0.1.50"
google-api-python-client = "^2.127.0"
ipykernel = "^6.29.4"
google-search-results = "^2.4.2"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
